reorganized field utils and fields
