+++ src/components/ConfigEditor/DatasourceConfigWizard/SchemaField.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { type useForm, Controller } from 'react-hook-form';
 import { useStyles2, Icon, Tooltip } from '@grafana/ui';
 import { resolveActiveOverride, formKey } from './config';
-import { isFieldRequired, buildValidationRules, parseItemErrors } from './fieldUtils';
-import { FieldInput } from './FieldInput';
+import { isFieldRequired, buildValidationRules, parseItemErrors } from './inputs/fieldUtils';
+import { FieldInput } from './inputs/FieldInput';
 import { getFieldStyles } from './styles';
 import type { FormValues } from './datasource';
 import type { ConfigField } from '../../../schema/schema';
