+++ src/components/ConfigEditor/DatasourceConfigWizard/inputs/ObjectArrayEditor.tsx
@@ -4,8 +4,8 @@ import { useStyles2, Button, Tooltip, Stack } from '@grafana/ui';
 import { type GrafanaTheme2 } from '@grafana/data';
 import { SECURE_FIELD_CONFIGURED, findActiveSecureOverride } from '../datasource';
 import { type ConfigField } from '../../../../schema/schema';
-import { FieldInput } from '../FieldInput';
-import { parseItemErrors } from '../fieldUtils';
+import { FieldInput } from './FieldInput';
+import { parseItemErrors } from './fieldUtils';
 import { type FormFieldRef } from './types';
 
 type Props = {
