+++ src/components/ConfigEditor/DatasourceConfigWizard/inputs/index.ts
@@ -4,3 +4,5 @@ export * from './IndexedPairEditor';
 export * from './ObjectArrayEditor';
 export * from './SecureFieldInput';
 export * from './StringArrayInput';
+export * from './fieldUtils';
+export { FieldInput, type FieldInputProps } from './FieldInput';
